typo refactor
