+++ maestro-app/src/test/java/bio/overture/maestro/app/infra/adapter/outbound/indexing/elasticsearch/FileCentricElasticSearchAdapterUnavaibilityTest.java
@@ -27,8 +27,8 @@ import bio.overture.maestro.domain.api.exception.FailureData;
 import bio.overture.maestro.domain.api.message.IndexResult;
 import bio.overture.maestro.domain.entities.indexing.FileCentricDocument;
 import bio.overture.maestro.domain.port.outbound.indexing.BatchIndexFilesCommand;
-import bio.overture.masestro.test.Fixture;
-import bio.overture.masestro.test.TestCategory;
+import bio.overture.maestro.test.Fixture;
+import bio.overture.maestro.test.TestCategory;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import java.io.IOException;
