+++ maestro-domain/src/main/java/bio/overture/maestro/test/Fixture.java
@@ -15,7 +15,7 @@
  *   along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package bio.overture.masestro.test;
+package bio.overture.maestro.test;
 
 import static bio.overture.maestro.domain.utility.StringUtilities.inputStreamToString;
